Hook up real Stock panel content.
