+++ client/header/activity-panel/slotfill/index.js
@@ -19,13 +19,12 @@ import ActivityPanelToggleBubble from '../toggle-bubble';
 import { H, Section } from '@woocommerce/components';
 import Inbox from '../panels/inbox';
 import Orders from '../panels/orders';
-import StockPanel from '../panels/stock';
+import Stock from '../panels/stock';
 import { recordEvent } from 'lib/tracks';
 import Reviews from '../panels/reviews';
 import WordPressNotices from '../wordpress-notices';
 import Tabs from './tabs';
 import Panel from './panel';
-import Stock from './panels/stock';
 
 const manageStock = getSetting( 'manageStock', 'no' );
 const reviewsEnabled = getSetting( 'reviewsEnabled', 'no' );
